% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagerank_kmeans.R
\name{kmeans}
\alias{kmeans}
\title{kmeans using R}
\usage{
kmeans(data, k, max_iter = 10000)
}
\arguments{
\item{data}{Sample points}

\item{k}{Number of clusters}

\item{max_iter}{Maximum Number Of Iterations}
}
\value{
Center and cluster of each sample
}
\description{
kmeans using R
}
\examples{
\dontrun{
kmeans(data, k=3, max_iter=5000)
}
}
