% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pagerankc}
\alias{pagerankc}
\title{pagerank using Rcpp}
\usage{
pagerankc(adj_matrix, alpha = 0.85, max_iter = 10000L)
}
\arguments{
\item{adj_matrix}{adjacency matrix}

\item{alpha}{Correction probability of transition probability matrix}

\item{max_iter}{Maximum Number Of Iterations}
}
\value{
Pagerank values for each node
}
\description{
pagerank using Rcpp
}
\examples{
\dontrun{
r <- pagerankc(adj_matrix, alpha=0.9, max_iter=5000)
}
}
