% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagerank_kmeans.R
\name{pagerankr}
\alias{pagerankr}
\title{Pagerank using R}
\usage{
pagerankr(adj_matrix, alpha = 0.85, max_iter = 10000)
}
\arguments{
\item{adj_matrix}{adjacency matrix}

\item{alpha}{Correction probability of transition probability matrix}

\item{max_iter}{Maximum Number Of Iterations}
}
\value{
Pagerank values for each node
}
\description{
Pagerank using R
}
\examples{
\dontrun{
r <- pagerankr(adj_matrix, alpha=0.9, max_iter=5000)
}
}
